+++ src/app/Http/Controllers/Admin/SocialMediaItemController.php
@@ -15,17 +15,23 @@ class SocialMediaItemController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', SocialMediaItem::class);
+
         $social_media = SocialMediaItem::orderBy('social_order')->get();
         return view('admin.social_media.index', compact('social_media'));
     }
 
     public function create()
     {
+        $this->authorize('create', SocialMediaItem::class);
+
         return view('admin.social_media.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', SocialMediaItem::class);
+
         $social_media = new SocialMediaItem();
         $data = $request->only($social_media->getFillable());
 
@@ -41,12 +47,16 @@ class SocialMediaItemController extends Controller
 
     public function edit($id)
     {
+        $this->authorize('update', SocialMediaItem::class);
+
         $social_media = SocialMediaItem::findOrFail($id);
         return view('admin.social_media.edit', compact('social_media'));
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('update', SocialMediaItem::class);
+
         $social_media = SocialMediaItem::findOrFail($id);
         $data = $request->only($social_media->getFillable());
 
@@ -62,6 +72,8 @@ class SocialMediaItemController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('delete', SocialMediaItem::class);
+
         $social_media = SocialMediaItem::findOrFail($id);
         $social_media->delete();
         return Redirect()->back()->with('success', 'Social Media Item is deleted successfully!');
